% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{map_params}
\alias{map_params}
\title{Map Paramters}
\usage{
map_params(params, .f, ..., use_names = T)
}
\arguments{
\item{params}{list generated by parse_arguments.}

\item{.f}{A function.}

\item{...}{Additional arguments passed on to the mapped function.}
}
\value{
A vector the same length as parameters.
}
\description{
Wrapper around map. Itterate over Parameters
}
