% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot_param.R
\name{mod_plot_param_ui}
\alias{mod_plot_param_ui}
\alias{mod_plot_param_server}
\title{mod_plot_param_ui and mod_plot_param_server}
\usage{
mod_plot_param_ui(id)

mod_plot_param_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
