% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_arguments}
\alias{parse_arguments}
\title{Parse Parameters}
\usage{
parse_arguments(params, score)
}
\arguments{
\item{params}{Parameters for previous modell training}

\item{score}{Score wich depand on params (same order)}
}
\value{
List of Variables
}
\description{
Parse Parameters and prepare them for futher computation
}
