% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plot_metric_point}
\alias{plot_metric_point}
\title{Visualize metric scores}
\usage{
plot_metric_point(df, param_name, score_name)
}
\arguments{
\item{df}{Dataframe with parameters and score.}

\item{param_name}{Name of parameter}

\item{score_name}{Name of score.}
}
\value{
ggplot element
}
\description{
Visualize metric scores
}
