% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_svg}
\alias{render_svg}
\title{Render a SVG}
\usage{
render_svg(
  g = NULL,
  height = 8,
  width = 8,
  scaling = 1,
  element_width = "100\%",
  element_height = "100\%",
  add_download_button = T,
  standalone = F,
  web_fonts =
    "https://fonts.googleapis.com/css2?family=Lato:wght@100;300;400;700;900&display=swap",
  bg = "transparent",
  font_family = "Lato, sans-serif",
  fix_rect = T,
  ...
)
}
\arguments{
\item{g}{A \code{ggplot} object or a \code{call} object}

\item{height}{Height of the SVG}

\item{width}{Width of the SVG}

\item{scaling}{scaling factor for the SVG}

\item{element_width}{width of the SVG element in any valid CSS unit}

\item{element_height}{height of the SVG element in any valid CSS unit}

\item{add_download_button}{if \code{TRUE} a download button will be added to the SVG}

\item{standalone}{if \code{TRUE} the SVG will be surrounded by an HTML document}

\item{web_fonts}{Fonts to be used for the SVG}

\item{bg}{Background color of the SVG}

\item{font_family}{Font family of the SVG}

\item{fix_rect}{if \code{TRUE} a fix rectangle elements will be applied
(solves problems with overlayed white rectanges in certain circumstances)}

\item{...}{Additional arguments passed to the SVG device}
}
\value{
A browsable SVG
}
\description{
Render a given ggplot object or a call object as a SVG
}
